Move Indices and Coordinates to starfish.constants
Test: python -v -m starfish 2>&1 | egrep '^import' yields no fat libraries.
