+++ starfish/constants.py
