+++ starfish/image/__init__.py
@@ -1,3 +1,2 @@
 from ._base import ImageBase
-from ._constants import Coordinates, Indices
 from ._stack import ImageStack
